move path_sampler
